% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_inel.R
\name{build_WRF_cache}
\alias{build_WRF_cache}
\title{Build WRF Cache}
\usage{
build_WRF_cache(wavelength, verbose = TRUE)
}
\arguments{
\item{wavelength}{Vector of wavelengths (nm)}

\item{verbose}{Print progress messages}
}
\value{
Invisible NULL (cache is stored in .WRF_cache environment)
}
\description{
Pre-computes a single base WRF matrix (phi_f = 1) for the given wavelength
grid. The actual quantum yield is applied at compute time inside
\code{sicf_analytical()}, exploiting the linearity
WRF(i,j; phi_f) = phi_f * WRF_base(i,j).
}
