% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_inel.R
\name{.compute_Ed_cached}
\alias{.compute_Ed_cached}
\title{Compute Ed with Caching}
\usage{
.compute_Ed_cached(
  wavelength,
  sunzen_deg,
  lat,
  lon,
  date_time,
  verbose = FALSE
)
}
\arguments{
\item{wavelength}{Vector of wavelengths (nm)}

\item{sunzen_deg}{Solar zenith angle (degrees)}

\item{lat}{Latitude (decimal degrees)}

\item{lon}{Longitude (decimal degrees)}

\item{date_time}{POSIXct datetime in UTC}

\item{verbose}{Print cache hit/miss messages}
}
\value{
List with Ed_0m (subsurface Ed) and E0_0m (subsurface scalar irradiance)
}
\description{
Wrapper around Gregg & Carder Ed calculation that caches results.
Cache key is based on all input parameters.
}
\keyword{internal}
