% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_inel.R
\name{.discretize_wrf_chl_vectorized}
\alias{.discretize_wrf_chl_vectorized}
\title{Discretize WRF - Fully Vectorized Alternative (Advanced)}
\usage{
.discretize_wrf_chl_vectorized(wavelength, quant_phi = 0.02)
}
\description{
This is an experimental fully-vectorized version that pre-computes
ALL wavelength combinations at once. Uses more memory but much faster.
}
\keyword{internal}
