% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_inel.R
\name{.ensure_WRF_cache}
\alias{.ensure_WRF_cache}
\title{Ensure WRF Cache is Initialized}
\usage{
.ensure_WRF_cache(wavelength, verbose = FALSE)
}
\arguments{
\item{wavelength}{Vector of wavelengths (nm)}

\item{verbose}{Print initialization messages}
}
\value{
Invisible NULL
}
\description{
Automatically initialize WRF cache if not already built for given wavelengths.
Called internally by sicf_analytical() - users don't need to call this directly.
}
\keyword{internal}
