% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gregg_carder.R
\name{.gc_atmodd}
\alias{.gc_atmodd}
\title{Compute spectral atmospheric transmittance and downwelling irradiance}
\usage{
.gc_atmodd(lam, theta, oza, ag, aw, sco3, p, wv, rh, am, wsm, ws, vis, Fo)
}
\description{
Compute spectral atmospheric transmittance and downwelling irradiance
}
\keyword{internal}
