% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gregg_carder.R
\name{.gc_navaer}
\alias{.gc_navaer}
\title{Compute Navy marine aerosol parameters}
\usage{
.gc_navaer(rh, am, wsm, ws, vis)
}
\arguments{
\item{rh}{Relative humidity (\%)}

\item{am}{Air mass type (1-10)}

\item{wsm}{Mean wind speed (m/s)}

\item{ws}{Current wind speed (m/s)}

\item{vis}{Visibility (km)}
}
\value{
List: beta, alpha, wa (single-scattering albedo), asymp
}
\description{
Compute Navy marine aerosol parameters
}
\keyword{internal}
