% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gregg_carder.R
\name{.gc_sfcrfl}
\alias{.gc_sfcrfl}
\title{Compute Fresnel surface reflectance}
\usage{
.gc_sfcrfl(theta, ws)
}
\arguments{
\item{theta}{Solar zenith angle (degrees)}

\item{ws}{Wind speed (m/s)}
}
\value{
List: rod (direct reflectance), ros (diffuse reflectance)
}
\description{
Compute Fresnel surface reflectance
}
\keyword{internal}
