% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gregg_carder.R
\name{.gc_sunang}
\alias{.gc_sunang}
\title{Compute solar zenith angle}
\usage{
.gc_sunang(iday, hr, xlon, ylat)
}
\arguments{
\item{iday}{Julian day (integer)}

\item{hr}{Time in decimal hours GMT}

\item{xlon}{Longitude, W negative (decimal degrees)}

\item{ylat}{Latitude, N positive (decimal degrees)}
}
\value{
Solar zenith angle in degrees
}
\description{
Compute solar zenith angle
}
\keyword{internal}
