% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.objective_function_registry}
\alias{.objective_function_registry}
\title{Objective Function Registry}
\usage{
.objective_function_registry
}
\description{
List of objective functions available to SABER
}
