% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_inel.R
\name{.plot_depth_profiles_sicf}
\alias{.plot_depth_profiles_sicf}
\title{Plot Depth Profiles of E0 and Fluorescence Attenuation}
\usage{
.plot_depth_profiles_sicf(
  wavelength,
  E0_surface,
  Kd,
  a_phy,
  WRF_matrix,
  phi_f,
  c_chl,
  a_dg_443,
  bb_p_550
)
}
\arguments{
\item{wavelength}{Vector of wavelengths (nm)}

\item{E0_surface}{Scalar irradiance at surface (W m^-2 nm^-1)}

\item{Kd}{Diffuse attenuation coefficients (m^-1)}

\item{a_phy}{Phytoplankton absorption (m^-1)}

\item{WRF_matrix}{Wavelength redistribution function matrix}

\item{phi_f}{Fluorescence quantum yield}

\item{c_chl}{Chlorophyll concentration for plot title}

\item{a_dg_443}{CDOM absorption for plot title}

\item{bb_p_550}{Particulate backscattering for plot title}
}
\value{
List with E0_z and Lf_z matrices, plus ggplot objects
}
\description{
Creates diagnostic plots showing how scalar irradiance and fluorescence
radiance decay with depth for different wavelengths.
}
\keyword{internal}
