% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_model.R
\name{forward_am03}
\alias{forward_am03}
\title{Albert & Mobley (2003) Forward Model}
\usage{
forward_am03(
  wavelength,
  iop,
  water_type = 2,
  theta_sun,
  theta_view,
  h_w = NULL,
  r_b = NULL
)
}
\arguments{
\item{wavelength}{vector of wavelengths \[nm\]}

\item{iop}{list with elements \code{a} and \code{bb}, same length as wavelength}

\item{water_type}{either 1 or 2 (default = 2)}

\item{theta_sun}{sun zenith angle \[degrees\]}

\item{theta_view}{sensor view angle \[degrees\]}

\item{h_w}{optional: water depth \[m\] (enables shallow mode)}

\item{r_b}{optional: bottom reflectance vector (same length as wavelength)}
}
\value{
numeric vector of subsurface Rrs
}
\description{
Computes Rrs below surface using analytical model from Albert & Mobley (2003)
}
\references{
Albert, A. and Mobley, C.D. (2003) ‘An analytical model for subsurface irradiance and remote sensing reflectance in deep and shallow case-2 waters’, Optics Express, 11(22), pp. 2873–2890. Available at: https://doi.org/10.1364/OE.11.002873.
}
