% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_inel.R
\name{get_Ed_cache_info}
\alias{get_Ed_cache_info}
\title{Get Ed Cache Info}
\usage{
get_Ed_cache_info()
}
\value{
List with number of cached entries
}
\description{
Return information about the current cache state.
}
