% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_inel.R
\name{get_WRF_cache_info}
\alias{get_WRF_cache_info}
\title{Get WRF Cache Info}
\usage{
get_WRF_cache_info()
}
\value{
List with cache details
}
\description{
Return information about the current WRF cache state.
}
