% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_objective.R
\name{lee99}
\alias{lee99}
\title{lee 1998 spectral error index}
\usage{
lee99(modelled, observed, wavelength)
}
\value{
the spectral error index of modelled vs observed values
}
\description{
lee 1998 spectral error index
}
\references{
Lee, Z. et al. (1999) ‘Hyperspectral remote sensing for shallow waters: 2 Deriving bottom depths and water properties by optimization’, Applied Optics, 38(18), p. 3831. Available at: https://doi.org/10.1364/AO.38.003831.
}
