% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_r_rs_b.R
\name{load_r_rs_b}
\alias{load_r_rs_b}
\title{Load bottom reflectance data}
\usage{
load_r_rs_b(wavelength, rrs_b)
}
\arguments{
\item{wavelength}{numeric vector}

\item{rrs_b}{matrix (columns = class names)}
}
\description{
Load bottom reflectance data
}
