% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_objective.R
\name{log_ll}
\alias{log_ll}
\title{log likelihood}
\usage{
log_ll(modelled, observed, sd, weights = NULL)
}
\value{
the likelihood of modeled vs observed values
}
\description{
log likelihood
}
