% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_iop_pure_water.R
\name{pure_water_iop}
\alias{pure_water_iop}
\title{pure_water_iop
compute pure water iop at requested wavelength}
\usage{
pure_water_iop(wavelength)
}
\arguments{
\item{wavelength}{vector of requested wavelength}

\item{water_type}{vector of requested wavelength}
}
\description{
pure_water_iop
compute pure water iop at requested wavelength
}
