% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_forward_model.R
\name{retrieve_r_rs_b_am03}
\alias{retrieve_r_rs_b_am03}
\title{Function to retrieve Rrs_b
\code{Rrs_Bottom <- (obs_rrs - Rrs_below_deep * (1 - Ars1 * exp(-zB * (Kd + kuW)))) / (Ars2 * exp(-zB * (Kd + kuB)))}}
\usage{
retrieve_r_rs_b_am03(
  wavelength,
  iop,
  r_rs,
  water_type = 2,
  theta_sun,
  theta_view,
  h_w
)
}
\description{
Function to retrieve Rrs_b
\code{Rrs_Bottom <- (obs_rrs - Rrs_below_deep * (1 - Ars1 * exp(-zB * (Kd + kuW)))) / (Ars2 * exp(-zB * (Kd + kuB)))}
}
