% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_r_rs_b.R
\name{select_benthic_classes}
\alias{select_benthic_classes}
\title{Select and load specific benthic classes}
\usage{
select_benthic_classes(classes)
}
\arguments{
\item{classes}{Character vector of class names to use (e.g., c("Mud_2019", "Eelgrass_2019", "Sand_2019"))}
}
\description{
Select and load specific benthic classes
}
