% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_optics.R
\name{snell_law}
\alias{snell_law}
\title{snell law, compute refracted angles and reflection}
\usage{
snell_law(theta_view, theta_sun)
}
\arguments{
\item{theta_view}{sensor zenith angle [deg]}

\item{theta_sun}{solar zenith angle [deg]}
}
\value{
list with \code{view_w}, \code{sun_w}, \code{rho_L}
}
\description{
snell law, compute refracted angles and reflection
}
