% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{validate_stan_installation}
\alias{validate_stan_installation}
\title{Validate Stan Model Compilation}
\usage{
validate_stan_installation()
}
\value{
Logical, TRUE if Stan is ready
}
\description{
Checks if Stan is properly installed and can compile models.
}
